# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Catalogs API Documentation",
        "description": "This is the API documentation for my Laravel application using Swagger/OpenAPI 3.0.",
        "contact": {
            "name": "API Support Team",
            "email": "ptryshaf@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/publisher": {
            "get": {
                "tags": [
                    "Publishers"
                ],
                "summary": "Get list of publisher",
                "operationId": "25315c4fe38fdbef40f7ae0b9c444420",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search keyword for publisher name or bio",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "John"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of publishers retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "List publishers successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "John Doe"
                                                    },
                                                    "address": {
                                                        "type": "string",
                                                        "example": "Lorem ipsum dolor sit amet"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "Lorem ipsum dolor sit amet"
                                                    },
                                                    "phone": {
                                                        "type": "string",
                                                        "example": "Lorem ipsum dolor sit amet"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2023-01-01T00:00:00.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2023-01-01T00:00:00.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Publishers"
                ],
                "summary": "Create a new publisher",
                "operationId": "59f2574b03355dea659b26eae515be7d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "address",
                                    "email",
                                    "phone"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Lorem ipsum dolor sit amet"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "Lorem ipsum dolor sit amet"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "Lorem ipsum dolor sit amet"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Publisher created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/publisher/{id}": {
            "get": {
                "tags": [
                    "Publishers"
                ],
                "summary": "Get detail of a publisher by ID",
                "operationId": "7cceb4560a45c3e9ed939e813c360f9e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Publisher ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Publisher retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "John Doe"
                                        },
                                        "address": {
                                            "type": "string",
                                            "example": "Lorem ipsum dolor sit amet"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "Lorem ipsum dolor sit amet"
                                        },
                                        "phone": {
                                            "type": "string",
                                            "example": "Lorem ipsum dolor sit amet"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-01-01T00:00:00.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-01-01T00:00:00.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "author not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Publishers"
                ],
                "summary": "Update publisher",
                "operationId": "d0b523327eb9fe89971b8d3ff628f642",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Publisher ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "address",
                                    "email",
                                    "phone"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Lorem ipsum dolor sit amet"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "Lorem ipsum dolor sit amet"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "Lorem ipsum dolor sit amet"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "publisher updated successfully"
                    },
                    "404": {
                        "description": "publisher not found"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Publishers"
                ],
                "summary": "Delete a publisher by ID",
                "operationId": "b05a7838e298fd6b1895b00b2f04127f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "publisher ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "publisher deleted successfully"
                    },
                    "404": {
                        "description": "publisher not found"
                    }
                }
            }
        }
    },
...
}